refactor: mouseDownTimer 커스텀 훅으로 분리
