+++ src/hooks/index.js
@@ -3,5 +3,13 @@ import useLocalStorage from './useLocalStorage';
 import useModal from './useModal';
 import useMultipleInput from './useMultipleInput';
 import useVirtualNumericKeyboard from './useVirtualNumericKeyboard';
+import useMousePressTimer from './useMousePressTimer';
 
-export { useInput, useLocalStorage, useModal, useMultipleInput, useVirtualNumericKeyboard };
+export {
+  useInput,
+  useLocalStorage,
+  useModal,
+  useMultipleInput,
+  useVirtualNumericKeyboard,
+  useMousePressTimer,
+};
